+++ docs/examples/calibration/plot_two_parameters.py
@@ -27,7 +27,7 @@ from gemseo.algos.parameter_space import ParameterSpace
 from gemseo.disciplines.analytic import AnalyticDiscipline
 from numpy import array
 
-from gemseo_calibration.scenario import CalibrationMeasure
+from gemseo_calibration.metrics.settings import CalibrationMetricSettings
 from gemseo_calibration.scenario import CalibrationScenario
 
 # %%
@@ -79,10 +79,13 @@ reference_data = reference_dataset.to_dict_of_arrays(False)
 # [CalibrationScenario][gemseo_calibration.scenario.CalibrationScenario]
 # to find the values of the parameters $a$ and $b$
 # which minimize a
-# [CalibrationMeasure][gemseo_calibration.measure.CalibrationMeasure]
+# [BaseCalibrationMetric][gemseo_calibration.metrics.base_calibration_metric.BaseCalibrationMetric]
 # taking into account the outputs $y$ and $z$:
-control_outputs = [CalibrationMeasure("y", "MSE"), CalibrationMeasure("z", "MSE")]
-calibration = CalibrationScenario(model, "x", control_outputs, prior)
+metric_settings = [
+    CalibrationMetricSettings(output_name="y", metric_name="MSE"),
+    CalibrationMetricSettings(output_name="z", metric_name="MSE"),
+]
+calibration = CalibrationScenario(model, "x", metric_settings, prior)
 calibration.execute(
     algo_name="NLOPT_COBYLA", reference_data=reference_data, max_iter=100
 )