+++ src/gemseo_calibration/calibrator.py
@@ -19,7 +19,6 @@ from __future__ import annotations
 import logging
 from typing import TYPE_CHECKING
 from typing import Any
-from typing import NamedTuple
 
 from gemseo.algos.design_space import DesignSpace
 from gemseo.algos.doe.custom_doe.custom_doe import CustomDOE
@@ -30,45 +29,23 @@ from gemseo.scenarios.doe_scenario import DOEScenario
 from numpy import array
 from numpy import hstack
 
-from gemseo_calibration.measures.factory import CalibrationMeasureFactory
+from gemseo_calibration.metrics.factory import CalibrationMetricFactory
+from gemseo_calibration.metrics.settings import CalibrationMetricSettings
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
     from collections.abc import Sequence
 
-    from gemseo.scenarios.base_scenario import BaseScenario
+    from gemseo.formulations.base_formulation_settings import BaseFormulationSettings
 
-    from gemseo_calibration.measure import CalibrationMeasure as CalibrationMeasure_
-    from gemseo_calibration.measure import DataType
-
-
-class CalibrationMeasure(NamedTuple):
-    """A calibration measure."""
-
-    output: str
-    """The name of the output."""
-
-    measure: str = "MSE"
-    """The name of the measure to compare the observed and simulated outputs."""
-
-    mesh: str | None = None
-    """The name of the irregular mesh associated with the output if any.
-
-    To be used when the output is a 1D function discretized over an irregular mesh.
-    """
-
-    weight: float | None = None
-    """The weight of this measure when the measure is an element of a collection.
-
-    The weight must be between 0 and 1. The sum of the weights of the elements in the
-    collection must be 1.
-    """
+    from gemseo_calibration.metrics.base_calibration_metric import BaseCalibrationMetric
+    from gemseo_calibration.metrics.base_calibration_metric import DataType
 
 
 class Calibrator(MDOScenarioAdapter):
     """A discipline with parameters calibrated from reference input-output data.
 
-    When it is executed from parameters values, it computes the calibration measure with
+    When it is executed from parameters values, it computes the calibration metric with
     respect to the reference data, provided through the
     [set_reference_data][gemseo_calibration.calibrator.Calibrator.set_reference_data]
     method.
@@ -80,80 +57,80 @@ class Calibrator(MDOScenarioAdapter):
         self,
         disciplines: Discipline | list[Discipline],
         input_names: str | Iterable[str],
-        control_outputs: CalibrationMeasure | Sequence[CalibrationMeasure],
+        metric_settings_models: CalibrationMetricSettings
+        | Sequence[CalibrationMetricSettings],
         parameter_names: str | Iterable[str],
-        formulation: str = "MDF",
-        **formulation_options: Any,
+        formulation_settings_model: BaseFormulationSettings | None = None,
+        **formulation_settings: Any,
     ) -> None:
         """
         Args:
             disciplines: The disciplines
                 whose parameters must be calibrated from the reference data.
             input_names: The names of the inputs to be considered for the calibration.
-            control_outputs: The names of the outputs used to calibrate the disciplines
-                with the name of the calibration measure and the corresponding weight
-                comprised between 0 and 1 (the weights must sum to 1).
-                When the output is a 1D function discretized over an irregular mesh,
+            metric_settings_models: A collection of calibration settings,
+                including the name of the observed output,
+                the name of the calibration metric
+                and the corresponding weight comprised between 0 and 1
+                (the weights must sum to 1).
+                When the output is a 1D function discretized over a mesh,
                 the name of the mesh can be provided.
-                E.g. `CalibrationMeasure(output="z", measure="MSE")`
-                `CalibrationMeasure(output="z", measure="MSE", weight=0.3)`
-                or `CalibrationMeasure(output="z", measure="MSE", mesh="z_mesh")`
-                Lastly, `CalibrationMeasure` can be imported
+                E.g. `CalibrationMetricSettings(output_name="z", metric_name="MSE")`
+                `CalibrationMetricSettings(output_name="z", metric_name="MSE", weight=0.3)`
+                or
+                `CalibrationMetricSettings(output_name="z", metric_name="MSE", mesh_name="z_mesh")`
+                Lastly, `CalibrationMetric` can be imported
                 from [gemseo_calibration.calibrator][gemseo_calibration.calibrator].
             parameter_names: The names of the parameters to be calibrated.
-            formulation: The name of a formulation
-                to manage the multidisciplinary coupling.
-            **formulation_options: The options of the formulation.
-        """  # noqa: D205,D212,D415
-        self.__measure_factory = CalibrationMeasureFactory()
+            formulation_settings_model: The MDO formulation settings
+                as a Pydantic model.
+                If ``None``, use ``**settings``.
+            **formulation_settings: The MDO formulation settings,
+                including the formulation name (use the keyword ``"formulation_name"``).
+                These arguments are ignored when ``settings_model`` is not ``None``.
+                If none and ``settings_model`` is ``None``,
+                the calibrator uses the default MDF formulation.
+        """  # noqa: D205,D212,D415,E501
+        self.__metric_factory = CalibrationMetricFactory()
         input_names = self.__to_iterable(input_names, str)
-        control_outputs = self.__to_iterable(control_outputs, CalibrationMeasure)
+        metric_settings_models = self.__to_iterable(
+            metric_settings_models, CalibrationMetricSettings
+        )
         parameter_names = self.__to_iterable(parameter_names, str)
         disciplines = self.__to_iterable(disciplines, Discipline)
-        control_output = control_outputs[0]
-        objective_name = control_output.output
-        mesh_name = control_output.mesh
         input_space = DesignSpace()
-        for name in input_names:
-            input_space.add_variable(name)
+        for input_name in input_names:
+            input_space.add_variable(input_name)
+
+        if (
+            formulation_settings_model is None
+            and "formulation_name" not in formulation_settings
+        ):
+            formulation_settings["formulation_name"] = "MDF"
 
         doe_scenario = DOEScenario(
             disciplines,
-            objective_name,
+            metric_settings_models[0].output_name,
             input_space,
-            formulation_name=formulation,
-            **formulation_options,
+            formulation_settings_model=formulation_settings_model,
+            **formulation_settings,
         )
-        if mesh_name:
+        if mesh_name := metric_settings_models[0].mesh_name:
             doe_scenario.add_observable(mesh_name)
 
-        self.__add_observables(control_outputs, doe_scenario)
+        for metric_settings_model in metric_settings_models[1:]:
+            doe_scenario.add_observable(metric_settings_model.output_name)
+            if mesh_name := metric_settings_model.mesh_name:
+                doe_scenario.add_observable(mesh_name)
 
         doe_scenario.set_algorithm(algo_name=CustomDOE.__name__)
 
-        self.__names_to_measures = {}
-        self.__measures = []
-        self.objective_name, output_names = self._add_measure(control_outputs)
+        self.__names_to_metrics = {}
+        self.__metrics = []
+        self.objective_name, output_names = self._add_metric(metric_settings_models)
         super().__init__(doe_scenario, parameter_names, output_names, name="Calibrator")
         self.__update_output_grammar()
         self.__reference_data = {}
-
-    @staticmethod
-    def __add_observables(
-        calibration_measures: Iterable[CalibrationMeasure], scenario: BaseScenario
-    ) -> None:
-        """Add observables to a scenario.
-
-        Args:
-            calibration_measures: The calibration measures to be added as observables.
-            scenario: The scenario.
-        """
-        for calibration_measure in calibration_measures:
-            output_name = calibration_measure.output
-            mesh_name = calibration_measure.mesh
-            scenario.add_observable(output_name)
-            if mesh_name:
-                scenario.add_observable(mesh_name)
 
     @staticmethod
     def __to_iterable(obj: Any, cls: type) -> Iterable[Any]:
@@ -174,16 +151,16 @@ class Calibrator(MDOScenarioAdapter):
         self.scenario.formulation.optimization_problem.reset()
 
     def __update_output_grammar(self) -> None:
-        """Redefine the output grammar from the names of the output measures.
+        """Redefine the output grammar from the names of the output metrics.
 
-        E.g. MSE(y,z) is the name of the MSE measure applied to the outputs y and z.
+        E.g. MSE(y,z) is the name of the MSE metric applied to the outputs y and z.
         """
         output_grammar = JSONGrammar("outputs")
-        output_grammar.update_from_names(self.__names_to_measures.keys())
+        output_grammar.update_from_names(self.__names_to_metrics.keys())
         self.output_grammar = output_grammar
 
     def set_reference_data(self, reference_data: DataType) -> None:
-        """Pass the reference data to the scenario and to the measures.
+        """Pass the reference data to the scenario and to the metrics.
 
         Args:
             reference_data: The reference data with which to compare the discipline.
@@ -201,8 +178,8 @@ class Calibrator(MDOScenarioAdapter):
                 for name in self.scenario.get_optim_variable_names()
             ]),
         )
-        for measure in self.__measures:
-            measure.set_reference_data(self.__reference_data)
+        for metric in self.__metrics:
+            metric.set_reference_data(self.__reference_data)
 
     def _execute(self) -> None:
         root_logger = logging.getLogger()
@@ -213,118 +190,130 @@ class Calibrator(MDOScenarioAdapter):
 
     def _post_run(self) -> None:
         model_dataset = self.scenario.to_dataset().to_dict_of_arrays(False)
-        for name, measure in self.__names_to_measures.items():
-            self.io.data[name] = array([measure.func(model_dataset)])
+        for name, metric in self.__names_to_metrics.items():
+            self.io.data[name] = array([metric.func(model_dataset)])
 
     @property
-    def maximize_objective_measure(self) -> bool:
-        """Whether to maximize the calibration measure related to the objectives."""
-        return self.__names_to_measures[self.objective_name].maximize
+    def maximize_objective_metric(self) -> bool:
+        """Whether to maximize the calibration metric related to the objectives."""
+        return self.__names_to_metrics[self.objective_name].maximize
 
-    def _add_measure(
+    def _add_metric(
         self,
-        control_outputs: CalibrationMeasure | Iterable[CalibrationMeasure],
+        metric_settings_models: CalibrationMetricSettings
+        | Iterable[CalibrationMetricSettings],
     ) -> tuple[str, list[str]]:
-        """Create a new calibration measure and add it to the outputs of the adapter.
+        """Create a new calibration metric and add it to the outputs of the adapter.
 
-        The purpose of this method is to decouple adding a measure from updating
+        The purpose of this method is to decouple adding a metric from updating
         the output grammar because during the call from __init__ the grammar instances
         are not yet existing before calling super().__init__.
 
         Args:
-            control_outputs: The names of the outputs used to calibrate the disciplines
-                with the name of the calibration measure and the corresponding weight
-                comprised between 0 and 1 (the weights must sum to 1).
-                When the output is a 1D function discretized over an irregular mesh,
-                the name of the mesh can be provided.
-                E.g. `CalibrationMeasure(output="z", measure="MSE")`
-                `CalibrationMeasure(output="z", measure="MSE", weight=0.3)`
-                or `CalibrationMeasure(output="z", measure="MSE", mesh="z_mesh")`
-                Lastly, `CalibrationMeasure` can be imported
-                from [gemseo_calibration.calibrator][gemseo_calibration.calibrator].
+            metric_settings_models: A collection of calibration settings.,
 
         Returns:
-            The name of the calibration measure applied to the outputs.
-        """
-        control_outputs = self.__update_weights(control_outputs)
+            The name of the calibration metric applied to the outputs.
+        """  # noqa: E501
+        metric_settings_models = self.__update_weights(metric_settings_models)
         name = ""
-        control_outputs = self.__to_control_outputs(control_outputs)
-        control_output = control_outputs[0]
-        measure, output_names = self.__create_measure(control_output)
-        self.__measures.append(measure)
-        maximize = measure.maximize
-        weight = control_output.weight
+        metric_settings_models = self.__to_metric_settings(metric_settings_models)
+        metric_settings_model = metric_settings_models[0]
+        metric, output_names = self.__create_metric(metric_settings_model)
+        self.__metrics.append(metric)
+        maximize = metric.maximize
+        weight = metric_settings_model.weight
         if weight != 1.0:
-            name += f"{weight}*{control_output.measure}[{measure.full_output_name}]"
-            measure = measure * weight
+            name += (
+                f"{weight}*{metric_settings_model.metric_name}"
+                f"[{metric.full_output_name}]"
+            )
+            metric = metric * weight
         else:
-            name += f"{control_output.measure}[{measure.full_output_name}]"
-        for control_output in control_outputs[1:]:
-            measure_, output_names_ = self.__create_measure(control_output)
-            self.__measures.append(measure_)
-            if measure_.maximize == maximize:
-                weight = control_output.weight
+            name += f"{metric_settings_model.metric_name}[{metric.full_output_name}]"
+        for metric_settings_model in metric_settings_models[1:]:
+            metric_, output_names_ = self.__create_metric(metric_settings_model)
+            self.__metrics.append(metric_)
+            if metric_.maximize == maximize:
+                weight = metric_settings_model.weight
             else:
-                weight = -control_output.weight
-            measure = measure + measure_ * weight
+                weight = -metric_settings_model.weight
+            metric = metric + metric_ * weight
             output_names.extend(output_names_)
-            name += f"+{weight}*{control_output.measure}[{measure_.full_output_name}]"
+            name += (
+                f"+{weight}*{metric_settings_model.metric_name}"
+                f"[{metric_.full_output_name}]"
+            )
 
-        measure.maximize = maximize
-        measure.name = name
-        self.__names_to_measures[name] = measure
+        metric.maximize = maximize
+        metric.name = name
+        self.__names_to_metrics[name] = metric
         return name, list(set(output_names))
 
     @staticmethod
-    def __to_control_outputs(
-        control_outputs: CalibrationMeasure | Iterable[CalibrationMeasure],
-    ) -> Iterable[CalibrationMeasure]:
-        """Force control output(s) to be an iterator of calibration measures.
+    def __to_metric_settings(
+        metric_settings_models: CalibrationMetricSettings
+        | Iterable[CalibrationMetricSettings],
+    ) -> Iterable[CalibrationMetricSettings]:
+        """Force calibration settings model(s) to be a collection.
 
         Args:
-            control_outputs: The control output(s).
-        """
-        if isinstance(control_outputs, CalibrationMeasure):
-            control_outputs = [control_outputs]
-        return control_outputs
-
-    def add_measure(
-        self,
-        control_outputs: CalibrationMeasure | Iterable[CalibrationMeasure],
-    ) -> tuple[str, list[str]]:
-        """Create a new calibration measure and add it to the outputs of the adapter.
-
-        Args:
-            control_outputs: The names of the outputs used to calibrate the disciplines
-                with the name of the calibration measure and the corresponding weight
-                comprised between 0 and 1 (the weights must sum to 1).
-                When the output is a 1D function discretized over an irregular mesh,
-                the name of the mesh can be provided.
-                E.g. ``CalibrationMeasure(output="z", measure="MSE")``
-                ``CalibrationMeasure(output="z", measure="MSE", weight=0.3)``
-                or ``CalibrationMeasure(output="z", measure="MSE", mesh="z_mesh")``
-                Lastly, ``CalibrationMeasure`` can be imported
-                from [gemseo_calibration.scenario][gemseo_calibration.scenario].
+            metric_settings_models: Either a calibration settings model
+                or a collection of calibration settings models.
 
         Returns:
-            The name of the calibration measure applied to the outputs.
+            A collection of calibration settings models.
         """
-        control_outputs = self.__to_control_outputs(control_outputs)
-        self.__add_observables(control_outputs, self.scenario)
-        return_values = self._add_measure(control_outputs)
+        if isinstance(metric_settings_models, CalibrationMetricSettings):
+            metric_settings_models = [metric_settings_models]
+        return metric_settings_models
+
+    def add_metric(
+        self,
+        metric_settings_models: CalibrationMetricSettings
+        | Iterable[CalibrationMetricSettings],
+    ) -> tuple[str, list[str]]:
+        """Create a new calibration metric and add it to the outputs of the adapter.
+
+        Args:
+            metric_settings_models: A collection of calibration settings,
+                including the name of the observed output,
+                the name of the calibration metric
+                and the corresponding weight comprised between 0 and 1
+                (the weights must sum to 1).
+                When the output is a 1D function discretized over a mesh,
+                the name of the mesh can be provided.
+                E.g. `CalibrationMetricSettings(output_name="z", metric_name="MSE")`
+                `CalibrationMetricSettings(output_name="z", metric_name="MSE", weight=0.3)`
+                or
+                `CalibrationMetricSettings(output_name="z", metric_name="MSE", mesh_name="z_mesh")`
+                from
+                [gemseo_calibration.metrics.settings][gemseo_calibration.metrics.settings].
+
+        Returns:
+            The name of the calibration metric applied to the outputs.
+        """  # noqa: E501
+        metric_settings_models = self.__to_metric_settings(metric_settings_models)
+        for metric_settings_model in metric_settings_models:
+            self.scenario.add_observable(metric_settings_model.output_name)
+            if mesh_name := metric_settings_model.mesh_name:
+                self.scenario.add_observable(mesh_name)
+
+        return_values = self._add_metric(metric_settings_models)
         self.__update_output_grammar()
         return return_values
 
+    @staticmethod
     def __update_weights(
-        self, control_outputs: Sequence[CalibrationMeasure]
-    ) -> Sequence[CalibrationMeasure]:
-        """Update the weights of the control outputs.
+        metric_settings_models: Sequence[CalibrationMetricSettings],
+    ) -> Sequence[CalibrationMetricSettings]:
+        """Update the weights of a collection of calibration settings.
 
         Args:
-            control_outputs: The control outputs.
+            metric_settings_models: A collection of calibration settings.
 
         Returns:
-            The updated control outputs.
+            The updated collection of calibration settings.
 
         Raises:
             ValueError: When a weight is outside [0, 1]
@@ -332,8 +321,8 @@ class Calibrator(MDOScenarioAdapter):
         """
         total_weight = 0
         missing_weight_indices = []
-        for index, control_output in enumerate(control_outputs):
-            weight = control_output.weight
+        for index, metric_settings_model in enumerate(metric_settings_models):
+            weight = metric_settings_model.weight
             if weight is None:
                 missing_weight_indices.append(index)
                 continue
@@ -342,13 +331,13 @@ class Calibrator(MDOScenarioAdapter):
                 msg = "The weight must be comprised between 0 and 1."
                 raise ValueError(msg)
 
-            total_weight += control_output.weight
+            total_weight += metric_settings_model.weight
 
         if not missing_weight_indices:
             if total_weight != 1:
                 msg = "The weights must sum to 1."
                 raise ValueError(msg)
-            return control_outputs
+            return metric_settings_models
 
         if total_weight >= 1:
             msg = "The weights must sum to 1."
@@ -356,31 +345,43 @@ class Calibrator(MDOScenarioAdapter):
 
         missing_weight = (1 - total_weight) / len(missing_weight_indices)
         for index in missing_weight_indices:
-            control_output = control_outputs[index]
-            control_outputs[index] = CalibrationMeasure(
-                output=control_output.output,
-                measure=control_output.measure,
-                mesh=control_output.mesh,
+            metric_setting = metric_settings_models[index]
+            metric_settings_models[index] = CalibrationMetricSettings(
+                output_name=metric_setting.output_name,
+                metric_name=metric_setting.metric_name,
+                mesh_name=metric_setting.mesh_name,
                 weight=missing_weight,
             )
 
-        return control_outputs
+        return metric_settings_models
 
-    def __create_measure(
-        self, control_output: CalibrationMeasure
-    ) -> tuple[CalibrationMeasure_, list[str]]:
-        if control_output.mesh:
-            measure = self.__measure_factory.create(
-                control_output.measure,
-                output_name=control_output.output,
-                mesh_name=control_output.mesh,
+    def __create_metric(
+        self, metric_settings_model: CalibrationMetricSettings
+    ) -> tuple[BaseCalibrationMetric, list[str]]:
+        """Create a calibration metric from settings.
+
+        Args:
+            metric_settings_model: The calibration settings.
+
+        Returns:
+            The calibration metric and the associated output name.
+        """
+        if metric_settings_model.mesh_name:
+            metric = self.__metric_factory.create(
+                metric_settings_model.metric_name,
+                output_name=metric_settings_model.output_name,
+                mesh_name=metric_settings_model.mesh_name,
             )
-            return measure, [control_output.output, control_output.mesh]
+            return metric, [
+                metric_settings_model.output_name,
+                metric_settings_model.mesh_name,
+            ]
 
-        measure = self.__measure_factory.create(
-            control_output.measure, output_name=control_output.output
+        metric = self.__metric_factory.create(
+            metric_settings_model.metric_name,
+            output_name=metric_settings_model.output_name,
         )
-        return measure, [control_output.output]
+        return metric, [metric_settings_model.output_name]
 
     @property
     def reference_data(self) -> DataType: