+++ src/gemseo_calibration/metrics/iae.py
@@ -12,6 +12,21 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program; if not, write to the Free Software Foundation,
 # Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-"""Measures of the consistency (or inconsistency) between two data sets."""
+"""Integrated absolute error between the model and reference output data."""
 
 from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from gemseo_calibration.metrics.base_integrated_metric import BaseIntegratedMetric
+
+if TYPE_CHECKING:
+    from gemseo.typing import RealArray
+
+
+class IAE(BaseIntegratedMetric):
+    """The integrated absolute error between the model and reference output data."""
+
+    @staticmethod
+    def _compare_data(data: RealArray, other_data: RealArray) -> RealArray:
+        return abs(data - other_data)