+++ src/gemseo_calibration/post_processor.py
@@ -26,7 +26,7 @@ from gemseo.post.base_post_settings import BasePostSettings
 if TYPE_CHECKING:
     from gemseo.algos.optimization_problem import OptimizationProblem
 
-    from gemseo_calibration.measure import DataType
+    from gemseo_calibration.metrics.base_calibration_metric import DataType
 
 T = TypeVar("T", bound=BasePostSettings)