+++ tests/test_calibrator.py
@@ -20,22 +20,25 @@ import pytest
 from numpy import array
 from numpy.testing import assert_equal
 
-from gemseo_calibration.calibrator import CalibrationMeasure
 from gemseo_calibration.calibrator import Calibrator
+from gemseo_calibration.metrics.settings import CalibrationMetricSettings
 
-CSTR_NAME = "0.5*MeasureCstr[y]+0.5*MeasureCstr[z]"
+CSTR_NAME = "0.5*MetricCstr[y]+0.5*MetricCstr[z]"
 
 
 @pytest.fixture
-def adapter(measure_factory, discipline) -> Calibrator:
-    """The calibration adapter to compute calibration measures from reference data."""
+def adapter(metric_factory, discipline) -> Calibrator:
+    """The calibration adapter to compute calibration metrics from reference data."""
     discipline = Calibrator(
-        discipline, ["x"], CalibrationMeasure("y", "MeasureObj"), ["a", "b"]
+        discipline,
+        ["x"],
+        CalibrationMetricSettings(output_name="y", metric_name="MetricObj"),
+        ["a", "b"],
     )
     discipline.default_input_data = {"a": array([0.5]), "b": array([0.5])}
-    discipline.add_measure([
-        CalibrationMeasure("y", "MeasureCstr"),
-        CalibrationMeasure("z", "MeasureCstr"),
+    discipline.add_metric([
+        CalibrationMetricSettings(output_name="y", metric_name="MetricCstr"),
+        CalibrationMetricSettings(output_name="z", metric_name="MetricCstr"),
     ])
     return discipline
 
@@ -43,14 +46,14 @@ def adapter(measure_factory, discipline) -> Calibrator:
 def test_init_objective(adapter):
     """Check the value of the objective name after initialization."""
     assert adapter.name == "Calibrator"
-    assert adapter.objective_name == "MeasureObj[y]"
+    assert adapter.objective_name == "MetricObj[y]"
 
 
-def test_init_measures(adapter):
-    """Check the value of the measures after initialization."""
-    names_to_measures = adapter._Calibrator__names_to_measures
-    assert str(names_to_measures["MeasureObj[y]"]) == "MeasureObj[y]"
-    assert str(names_to_measures[CSTR_NAME]) == CSTR_NAME
+def test_init_metrics(adapter):
+    """Check the value of the metrics after initialization."""
+    names_to_metrics = adapter._Calibrator__names_to_metrics
+    assert str(names_to_metrics["MetricObj[y]"]) == "MetricObj[y]"
+    assert str(names_to_metrics[CSTR_NAME]) == CSTR_NAME
 
 
 def test_init_data(adapter):
@@ -62,14 +65,14 @@ def test_init_grammars(adapter):
     """Check the value of the input and output grammars after initialization."""
     assert sorted(adapter.io.input_grammar.names) == ["a", "b"]
     assert sorted(adapter.io.output_grammar.names) == [
-        "0.5*MeasureCstr[y]+0.5*MeasureCstr[z]",
-        "MeasureObj[y]",
+        "0.5*MetricCstr[y]+0.5*MetricCstr[z]",
+        "MetricObj[y]",
     ]
 
 
 def test_maximize(adapter):
-    """Check the property 'maximize_objective_measure' after initialization."""
-    assert adapter.maximize_objective_measure is True
+    """Check the property 'maximize_objective_metric' after initialization."""
+    assert adapter.maximize_objective_metric is True
 
 
 def test_set_reference_data(adapter, reference_data):
@@ -80,8 +83,8 @@ def test_set_reference_data(adapter, reference_data):
         2,
         1,
     )
-    for measure in adapter._Calibrator__measures:
-        assert len(measure._reference_data) == 2
+    for metric in adapter._Calibrator__metrics:
+        assert len(metric._reference_data) == 2
 
     assert_equal(adapter.reference_data, reference_data)
 
@@ -90,7 +93,7 @@ def test_execute_default(adapter, reference_data):
     """Check the execution of the Calibrator with default input data."""
     adapter.set_reference_data(reference_data)
     adapter.execute()
-    assert adapter.io.data["MeasureObj[y]"][0] == 0.25
+    assert adapter.io.data["MetricObj[y]"][0] == 0.25
     assert adapter.io.data[CSTR_NAME][0] == 0.5
 
 
@@ -98,8 +101,8 @@ def test_execute(adapter, reference_data):
     """Check the execution of the Calibrator with custom input data."""
     adapter.set_reference_data(reference_data)
     adapter.execute({"a": array([0.25])})
-    assert adapter.io.data["MeasureObj[y]"][0] == 0.125
+    assert adapter.io.data["MetricObj[y]"][0] == 0.125
     assert adapter.io.data[CSTR_NAME][0] == 0.25
     adapter.execute({"a": array([0.75])})
-    assert adapter.io.data["MeasureObj[y]"][0] == 0.375
+    assert adapter.io.data["MetricObj[y]"][0] == 0.375
     assert adapter.io.data[CSTR_NAME][0] == 0.75